+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 [unreleased diff]
 
+### Fixed
+
+- Inconsistent padding for finishing up message in modern style.
+
 ## [2.1.2] - 2024-09-10
 
 [2.1.2 diff]