+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Fixed
 
 - Inconsistent padding for finishing up message in modern style.
+- Inconsistent padding for audio controls.
 
 ## [2.1.2] - 2024-09-10